+++ src/collection.js
@@ -1,0 +1,4 @@
+import _ from 'lodash/fp'
+
+export const flowMap = (...fns) => _.map(_.flow(...fns))
+export let findApply = _.curry((f, arr) => f(_.find(f, arr)))